backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
bidict==0.23.1
bleach==6.3.0
blinker==1.9.0
boto3==1.42.17
botocore==1.42.17
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
cramjam==2.11.0
cryptography==46.0.3
cysignals==1.12.6
Cython==3.2.3
DataProfiler==0.13.4
datasketches==5.2.0
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.8.0
docutils==0.22.4
executing==2.2.1
fastavro==1.12.1
fastjsonschema==2.21.2
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.5
fqdn==1.5.1
gevent==25.9.1
gmpy2==2.2.2
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.15.1
HLL==2.4.0
httpcore==1.0.9
httpx==0.28.1
idna==3.1
invoke==2.2.1
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jzpack==0.2.0
lark==1.3.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
memory_allocator==0.1.4
mistune==3.2.0
modcall==0.1.0
mpmath==1.3.0
msgpack==1.1.2
namespace==0.1.4
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.10.6
networkx==3.6.1
neuroglyph==2.3.1
numpy==1.26.2
opencv-python==4.8.1.78
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
paramiko==4.0.0
parso==0.8.5
passagemath-categories==10.6.42
passagemath-environment==10.6.42
passagemath-glpk==10.6.42
passagemath-modules==10.6.42
passagemath-objects==10.6.42
passagemath-polyhedra==10.6.42
passagemath-ppl==0.8.10.2
pathlib==1.0.1
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.5.1
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psutil==7.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
pyarrow==22.0.0
pycparser==2.23
pycryptodome==3.19.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyNaCl==1.6.1
pyspark==4.1.0
pyspark-analyzer==5.0.2
python-dateutil==2.9.0.post0
python-engineio==4.13.0
python-json-logger==4.0.0
python-snappy==0.7.3
python-socketio==5.10.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
redis==5.0.4
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
s3transfer==0.16.0
scipy==1.16.3
Send2Trash==1.8.3
simple-websocket==1.1.0
six==1.17.0
socketio==0.2.1
soupsieve==2.8.1
stack-data==0.6.3
starlette==0.50.0
statistics==1.0.3.5
StegoImageX==0.0.1
storage==0.0.4.3
terminado==0.18.1
tinycss2==1.4.0
tornado==6.5.4
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.2
uuid==1.30
versioneer==0.29
voila==0.5.11
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.0.1
wheel==0.45.1
wsproto==1.3.2
zope.event==6.1
zope.interface==8.1.1
zstandard==0.25.0

